refactor: полный аудит и оптимизация кода (26 проблем)
Исправлены проблемы выявленные в ходе аудита кода:

HIGH SEVERITY (4 проблемы):
- Исправлена проверка границ в can_rotate_with_wall_kick() (#3)
- Упрощена обработка ошибок в Drop реализациях (#7)
- Улучшена валидация путей с symlink_metadata() (#9)
- Добавлена константа MAX_SAFE_F32_FOR_U32 (#25)

MEDIUM SEVERITY (10 проблем):
- Приватные поля с валидирующими сеттерами в GameStats (#2)
- Добавлено логирование ошибок в handle_input() (#14)
- Расширена валидация Unicode с логированием (#18)
- Добавлена обработка ошибок confy с backup (#23)
- Обновлена документация SOLID в GameState (#12)
- Добавлены security-focused тесты (#10)

LOW SEVERITY (12 проблем):
- Оптимизировано кэширование строк с capacity (#4)
- Упрощён sanitize_player_name через matches! (#6)
- Возврат битовой маски в find_filled_lines() (#11)
- Удалены избыточные трейты в cycle.rs (#13)
- Добавлен #[inline] для check_block_collision() (#15)
- Расширена документация API в exports.rs (#16)
- Уменьшен размер Tetromino до 32 байт (#20)
- Расширены бенчмарки (#21)
- Удалены #[allow(dead_code)] (#24)
- Расширена документация безопасности (#26)

Документация:
- Обновлена TESTS_REGISTRY.md (1092 теста, 66 файлов)
- Обновлена ARCHITECTURE.md (оценка 9.2/10)
- Добавлен re-export sanitize_player_name

Тесты:
- Добавлены тесты для security fixes
- Все 1092 теста проходят успешно

Clippy: без предупреждений
Fmt: код отформатирован

+++ src/game/view.rs
@@ -409,7 +409,7 @@ impl<'a> GameView<'a> {
                 canvas.draw_strs(
                     &[SHAPE_STR],
                     (x as u16, y as u16),
-                    SHAPE_COLORS[self.curr_shape.fg],
+                    SHAPE_COLORS[self.curr_shape.fg as usize],
                     &Reset,
                 );
             }
